feat(web): affiche le salon secret aux utilisateurs présent dans le salon secret
